Reverted Generator and Help command to previous (hopefully working) states
